hands-on-projects/project-08-data-pipeline/app/pipeline-producer/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
kafka-python==2.0.2
